Upgrade to OnTopic 4.5.0
This includes a number of bug fixes and backend feature improvements. While these are useful for developers, and help improve reliability of the site, they shouldn't have much impact on the day-to-day experience.

Most notably, this includes a lot of new SQL scripts which make it easier to query and debug the database by simplifying tree traversal operations as well as filtering attributes.

This also fixes a bug which prevented the editor from deleting existing attributes. They could be overwritten with a new value, but if the value was removed entirely then that attribute might not get saved.