Upgrade to OnTopic Editor 4.5.0
This includes a couple of new attribute types, `InstructionAttribute` and `IncomingRelationshipAttribute`, which make it easier to understand comple data models involving interdependent relationships, such as those used on GoldSim's `ApplicationPage`.

This also includes a number of UI improvements, including "sticky" navigation, basic responsiveness, improved navigation of the relationship tree view, and the ability to see all validation errors reported in the tabs (whereas, previously, this only worked for validation errors from `input` boxes).

Critically, this takes advantage of a bug fix in the underlying OnTopic library, which resolves an issue where, under some circumstance, an existing attribute value could not be deleted, only overwritte with another non-zero-length value.

Finally, this includes a number of performance improvements which reduce the size of the CSS, and reduce how many blocking tasks are involved on each page load by deferring most (but not all) scripts.