Deferred HTML script
The recent update to OnTopic Editor 4.5.0 defers the loading of the JavaScript dependencies in order to improve performance. That means that jQuery isn't immediately available.

To accommodate this, the loading of the `HtmlViewComponent`'s view (which overrides the corresponding view from OnTopic Editor in order to provide customizations) needs to also be deferred.

We already did this to most GoldSim code as part of the Lighthouse performance updates. This apparently got missed. Oops!