Added Table of Contents support
Added optional support for adding a table of contents to any `ContentList` or `LinkedList` view, via `TableOfContents`. This can now be configured directly in the editor via the _Enable Table of Contents_ attribute (`IsIndexed`). Optionally, the label can also be set via _Table of Contents (Label)_ (`IndexLabel`). With this, I've also exposed `TableOfContents` to the `LinkedList` view so it's additionally available there.

This includes the new smooth scrolling support as well (fc20ad5).