Added sticky navigation for main menu
I had also experimented with making the page-level navigation sticky (as permitted by 4e1c6c9), but this introduced ambiguities since a) the white background blended in with the content area, and b) the relationship between the main navigation and the page-level navigation became obscured. We can revisit that later, if there's interest. But, for now, this allows the main menu to stick to the top.