Implemented new `GetDouble()` extension method
Previously, the OnTopic library had extension methods for retrieving topic attributes via e.g. `GetInteger()`, but not `GetDouble()`. Since the `PaymentsController` relies on a double value for the `InvoiceAmount`, this required custom conversion of the string value. With OnTopic 4.5, however, we now have support for `GetDouble()`, which simplifies this logic, and centralizes the conversion code.