Reordered Zurb Foundation scripts
Since the scripts are marked as `defer`, they will be executed in the order they appear on the page. Given that, I've updated them to reflect the priority order of content on the page, with the menu coming first. This likely won't make a noticeable difference, but it's a small optimization that's easy to make.