Updated `Html` view component override for OnTopic 4.5.0
With OnTopic 4.5.0, the view for the `HtmlViewComponent` is greatly simplified. This is because the jQuery to deduplicate IDs within the iFrame is no longer needed—and, in fact, will cause an error. Given that, we can also remove the `DOMContentLoader` event listener for deferring the script, as the `CKEDITOR` itself is not deferred. That also eliminates the need for wrapping the configuration in the `setEditorInstance()` function.