Applied offset to `html` element for `#anchor` links
In a previous update, I made the header sticky (at least on desktop). This works well, but means that when the browser scrolls to an anchor link, it hides the header under the navigation. By adding a `scroll-padding-top` offset based on the height of the header, we mitigate that.

This won't work for all scenarios, such as the accordions, since they rely on JavaScript for positioning, but it will address the bulk of scenarios across the site.