Enable `deep-link-smudge` feature for accordions
A previous fix resolved an issue where the header of the accordion slid under the sticky navigation when a user clicked on it (5b0442e). That didn't fix same scenario when a user navigated to an `id` using a `#hash` in the URL. For most scenario's, that was addressed by another commit, which applied a `scroll-padding-top` (9287ab8). That still failed for the accordions, however, due to the fact that the `#hash` references the content panel, not the `-label`.

To mitigate that, we can implement the `data-deep-link-smudge` feature in the Foundation Accordion control. The Accordion control also has a `data-deep-link-smudge-offset` option which we could use to account for the header, but the out-of-the-box behavior actually seems to provide enough of a buffer that this isn't an issue. As such, we're not including that here.