Rolled back `data-deep-link-smudge` implementation
The `data-deep-link-smudge` behavior previously enabled (151c237) does not operate as expected. It ensures that the top of the accordion _panel_ is visible, not the top of the current accordion _item_. That's not what we want, as it can result in the current item being pushed offscreen entirely when working with lengthy accordions.

Given this, I've returned to the default behavior, which results in the title being hidden behind the sticky navigation. This isn't optimal, but it at least ensures that the current item remains visible.

In addition, I've submitted a feature request to Zurb (foundation/foundation-sites#12176) so that, hopefully, we might see the desired behavior in a future release of Foundation Sites.