Set collection properties to be read only
As a best practice, collection properties should be read only so that they can't be set to new collection references. The collections can still be modified, obviously, they just can't be replaced. This necessitates that the read only properties be initialized with a new collection; this is acheived using the new C# 9.0 `new()` shortcut.

This resolves CA2227.