Suppress CA1724
CA1724 prevents classes from having the same name as namespaces. That's usually good guidance. In this case, as it can help prevent ambiguity and potential bugs. Unfortunately, as this name inherits from an OnTopic content type descriptor, that's not practical here while maintaining the convention unless we make a major refactoring. As this isn't introducing any actual bugs, and we don't expect there to be much ambiguity in context, I'm instead suppressing it in the source.