Implemented C# 9.0 implicit `new()`
In a previous commit (8991dc6), I converted collection properties to be read-only using the new implicit `new()`. In this update, I expand the use of `new()` to other areas. This includes instantiating properties of other object types via the constructor, retrofitting existing read-only collection properties, and applying it to propery initializers in calling code. This reduces the code size a bit, and aids in readability.