Reordered routes to fix forms
Our routing rules remain pretty fussy due to the fact that routes used for e.g. forms or tag helpers will not utilize the `DynamicRouteValueTransformer` routes, and thus you can route to a page via the URL, but any forms or tag helpers pointing to that page won't use the same URL. The exact mechanics of this changed slightly between ASP.NET Core 3.x and ASP.NET Core 5.0, so this update rearranges the order to mitigate that issue.