Updated to preview of OnTopic Library, Editor 4.6.0
The official release of OnTopic 4.6.0 is still forthcoming, but the basic updates are expected to be stable at this point. As such, we're able to dogfood the preview today. We'll obviously closely monitor the results in Application Insights, though we're not expecting any issues since this is mostly adoption of C# 9.0 syntactical updates.