Applied SemVer constraints to NuGet packages
We don't want to upgrade to EPPlus 5+, which is now a paid version. Likewise, we don't (yet) want to upgrade to .NET 5+ libraries, as they aren't comparible with an ASP.NET Core 3.x project. By using the SemVer constraints, NuGet will still notify us of updates to each which are under 5.x, but won't prompt to update to 5+ libraries. This will help make it easier to filter through available NuGet package updates.