Use out-of-the-box `ContentListTopicViewModel`
Previously, GoldSim replaced the out-of-the-box `ContentListTopicViewModel` with a customized version which included the `IsIndexed` and `IndexLabel` properties, thus supporting an optional table of contents to be added to any content list or derivative.

These have now been added into the preview version of OnTopic 4.6.0, however, and no longer need to be explicitly extended on GoldSim. As such, this can be removed entirely, and all references can fall back to the out-of-the-box version from `OnTopic.ViewModels`.