Throw `InvalidOperationException` for configuration issues
For non-public libraries we have historically thrown the general `Exception` for errors intended to be caught at design time, and in place to provide guidance to developers. These are not intended to be caught, so there isn't a concern with them being general. That said, this isn't a best practice and is easy to fix, so they're being replaced with `InvalidOperationException`s. This resolves CA2201.