Updated `[Relationship()]` to `[Collection()]`
In OnTopic 5.0.0, the `[Relationship()]` attribute has been renamed to `[Collection()]`, and the `RelationshipType` enum it accepts as an argument has been renamed to `CollectionType`. This is a less ambiguous and more descriptive label given the purpose it serves (which is to map the collection on the view model to a collection on the topic class).