Updated `[Follow()]` to `[Include()]`
In OnTopic 5.0.0, the `[Follow()]` attribute has been renamed to `[Include()]`, and the `Relationships` enum it accepts as an argument has been renamed to `AssociationTypes`. This is a less ambiguous and more descriptive label given the purpose it serves (which is to determine if properties referencing topic associations should be mapped).