Update view models to `record` types
In OnTopic 5.0.0, it is preferred that view models be implemented as C# 9.0 `record` types, which are immutable, to reinforce that they should be light-weight, read-only, value-like objects.

This introduces some downstream impacts due to controllers or components mapping a view model and then modifying its properties, since those properties can only be set via initialization. That will be addressed in subsequent commits.