Updated mapping logic to account for records
Previously, we mapped the view model first, then set conditional properties. That isn't possible with records since properties are `init` only. Instead, we must create the view model first with those conditional properties, and then map the remaining properties. This works because the mapping service relies on reflection, which is able to set values after `init`.