Updated `FormsController` to operate off of `record` types
This isn't an especially elegant solution. Binding models are nice because, previously, they operated off of an interface, so there was no need for a common base class. Generic constraints, however, cannot specify `record`—and, without that, the `with` keyword is unavailable, since it is restricted to `record` types. Later, in C# 10, this won't be an issue, as Microsoft will be extending the `with` syntax to classes. Until then, we need to specify a concrete `record` type as the type constraint.

Fortunately, for today's forms, all form view models ultimately derive from `CoreContact`, so we can safely use that as the constraint. We may need to refactor this at a later date, however, if we end up with a form that doesn't operate off of `CoreContact`.