Update the `IndexedList` view to work with view models
The `IndexedList` view was always a bit of a hack. The preferred approach, as of OnTopic 4.6.0, is to instead set the new `IsIndexed` attribute, which allows a table of contents to be added to _any_ `ContentList` view, instead of having a specific view just for that feature. That said, for backward compatibility, we should ideally maintain the `IndexedList` for any pages that have it set as their `View`.

To do this, we can no longer set the `IsVisited` property since it's an `init` only property on a `record`. Instead, we need to reassign the `Model` to a new variable, and explicitly pass the resulting value to the `ContentList` view.