Updated `NavigationViewModel` to operate off of `CurrentWebPath`
With OnTopic 5.0.0 Beta 4, the `NavigationViewModel<T>` has been updated to use `CurrentWebPath` instead of `CurrentKey`, and `NavigationTopicViewModel` has been updated to use `IsSelected(webPath)` instead of `IsSelected(uniqueKey)`. This reduces the number of properties required on the `NavigationTopicViewModel`.

It is also a more logical comparison since we're working with paths—and, indeed, several of GoldSim's view components fellback to `HttpContext.Request.Path` if `CurrentTopic.GetUniqueKey()` couldn't be identified (e.g., if there wasn't a `CurrentTopic` due to a 404). It's doubtful that scenario would have worked as expected previously, but it should work now.