Introduced new `CoreNavigationTopicViewModel`
The new `CoreNavigationTopicViewModel<T>` centralizes common elements between GoldSim's `NavigationTopicViewModel` and `TrackedNavigationTopicViewModel` classes.

It also no longer derives from `PageViewModel`, but instead adds the `Title` and `WebPath` requirements of `INavigationTopicViewModel<T>` directly. This makes for a smaller, simpler model.

More importantly, it also allows for the `NavigationTopicViewModel`s to be classes instead of records. Normally, records are preferred for view models. But the `TrackedNavigationTopicViewModel` includes an `IsVisited` property that needs to be writable on a per request basis in order to satisfy the needs of the navigation tracking.

Technically, this could be done with records by using `with {}` to make a shallow copy. But that's pretty wasteful here since the mapping library would have just created a record for every navigation item, and then e.g. the `LessonListViewComponent` would need to recreate every one of those with a new `IsVisited` property. Since `NavigationTopicViewModel` is already treated quite differently from other view models, it makes more sense to handle it as a class to address this unique need.

(Note that the `NavigationTopicViewModel` itself could easily be a record. But in order to share logic between it and `TrackedNavigationTopicViewModel` via the new `CoreNavigationTopicViewModel`, all three need to be classes.)