Reintroduced `EditorViewModelLookupService`
In OnTopic Editor 5.0.0 Beta 3, the `EditorViewModelLookupService` had been replaced with the out-of-the-box `DynamicTopicViewModelLookupService`. This worked fine, but it paints with a broad brush—and, in fact, would grab all of the view models from the `GoldSimTopicViewModelLookupService` as well as the `TopicViewModelLookupService`.

There's no problem using that if that's preferred, but OnTopic Editor 5.0.0 Beta 4 now reintroduces the `EditorViewModelLookupService`. The new version is now dynamic, so it'll find view models from any attribute type plugin assemblies. But it limits those to view models that derive from the OnTopic Editor's `ContentTypeDescriptorViewModel` or `AttributeDescriptorViewModel`—i.e., only the view models that are used by the OnTopic Editor. That provides a much smaller list, and makes more sense when combined as part of the `CompositeTypeLookupService`.