Fixed a bug in how `IsSelected()` was being evaluated
The `IsSelected()` was appending a `:` to both the `webPath` and the `WebPath`. This was useful when this used `UniqueKey` because it prevented a partial key match (e.g., `Root:Web:Pro` matching with `Root:WebProducts`. Unlike `UniqueKey`, `WebPath` returns a trailing delimiter—which is a `/` for `WebPath`. But as the `webPath` parameter might be derived from `HttpContext`, it should add a `/` to be safe.

This resolves an issue where `IsSelected()` would correctly match with an _exact_ match, but not a _partial_ match, thus failing to highlight parent nodes of the current topic.