Explicitly reset `area` when calling `RedirectToAction()` within area
Previously, the area wasn't set, which means it defaulted to the current area. That's fine as long as it's using the non-area route; but in .NET 5.0 it's _correctly_ using the area route, which means this points to `/Forms/Redirect/ID`. This can be fixed by resetting the `area` attribute in the dictionary.