Prevent nested paragraphs from extending parent's bottom margin
The editor likes to wrap text within paragraph tags. When these are placed within a _Note_ or a _Help_ link, that means the paragraph's bottom margin is added to the parent's, resulting in an inconsistent experience. This can be mitigated by removing the `margin-bottom` on the last paragraph within a _Note_ or _Help_ block. By only applying it to the last paragraph, we ensure there remain breaks between paragraphs, in case a block has multiple paragraphs. That's unexpected, but there's no reason to prevent that possibility.