Float block icons to the left
When using the `::before` pseudo-class to inject content before the e.g., `blockquote`, as is done for _Help_ and _Note_, any paragraphs will get placed _below_ the original position of the icon. (Oddly, inline text isn't affected by this, but paragraphs are.) This results in very different margins and layout if the block contains a paragraph. That isn't the expected markup, but it happens transparently when applying a _Note_ or _Help_ to an existing paragraph. To mitigate that, I'm applying the `float: left` to these icons, so the text doesn't attempt to wrap around them. With the _Note_ text, this also required adjusting the `top`, since it no longer needed to account for the space it had originally taken up, which was artificially increasing the margin.