Reset `margin-bottom` for nested lists
The margin is reduced on the last list item in a list. This is intended to prevent the list from extending the margin of its containing block. But, in doing so, it introduces presentation issues with nested lists. This helps mitigate this.

(Technically, this introduces the original problem if there is a nested list on the last of the parent items, but this looks better than the alternative. If it becomes a concern, we can take a more sophisticated approach. This is meant as a quick fix to a corner case.)