Updated `App_Initializing` content
The `App_Initializing.htm` had not been updated since the major cleanup of the stylesheets last year. As a result, it wasn't rendering correctly. This fixes the path to the stylesheet, as well as the markup as class structure of the content. While I was at it, I also added a byline.