Removed Entity Framework and Code Generation packages
These were injected by Visual Studio when creating a view off of a model. This is a rare use case, and it's unclear if current versions of Visual Studio still maintain that dependency. If they do, they can reinstall these dependencies the next time we find ourselves needing this functionality. As it's rarely used, however, it's preferred to remove these.