Update to OnTopic Editor 5.1.0 Beta 1 w/ bug fix
The previous update of OnTopic Editor 5.1.0 Beta 1 (8171ec7) included a major bug which prevented pages with unchecked radio buttons not to pass model validation. This issue is resolved. Due to an error in the build process, the version wasn't incremented, so we're using the pull request build associated with the release candidate as a shortcut so we don't need to make a faux commit. We'll be updating to the release version soon enough.