Introduced the `AssociatedTopicViewModel`
The `AssociatedTopicViewModel` is intended to capture the most commonly referenced properties used when cross-referencing other topics on, notably, the application pages. The goal is to rely on `[MapAs()]` so we can map a more targeted association type, whereas currently we map the entire target page (e.g., `ApplicationPageTopicViewModel`, `ExampleApplicationTopicViewModel`), even though we only use a small portion of it.

This obviously includes implements `ICardViewModel` (for the indexes) as well as `INavigableTopicViewModel` (for basic lists). It doesn't strictly need `ICoreTopicViewModel`, though this offers a lot more flexibility (such as the ability to use `TopicViewModelCollection`) with minimal extra cost (it required adding `ContentType`).

The `LearnMore` is needed to support e.g. white papers, and `LastModified` is needed to control the sort order. These are a bit extraneous for a number of scenarios, but are fast to map. If we want, in the future, we can further divide this up into a couple of different types, but as a first attempt this should provide a lot of benefit.