Updated `Slideshow` to account for `TransitionEffect`
It turns out that the `SlideshowTopicViewModel.TransitionEffect` was never exposed to the view. This may not be a big issue; this doesn't seem to work as advertised. Given that, we should evaluate removing it in favor of e.g. `AnimateIn` and `AnimateOut`. But in the meanwhile, it should be properly wired up.