Inject implicit space around breaks within a list
The WYSIWYG editor we use (CKEditor) doesn't (easily) allow paragraphs—or other block-formatted elements—to be added to a list. This is because a soft break (shift+enter) injects a `<br>`, but a hard-break (enter) inserts a new _list item_. There's no great way around this.

To mitigate this, we're introducing a CSS hack in which the `<br>` is treated as a block element by giving it content. I prefer to avoid hacks, and this clearly violates the _intention_ of the `<br>` tag, but it seems to be supported in major modern browsers, and it allows the common use of shift+enter to add the expected level of padding.