Increased maximum width of the courses area
Normally, the maximum width of the `grid-container` is set to 960px. With the courses area, we want to keep the white `panel` the same width as on the main site, but carve out room for the left-hand navigation. To accomplish this, I'm setting the `grid-container` to 1152px, which results in the main content area being ≈960px.

Because this stylesheet is shared between the courses, course, and lesson pages, this actually results in the main white area being _wider_ than the content area on the main site. This is inconsistent. But I think this actually works better because otherwise, the width of the site and the placement of the white box jumps when navigating between the course overview and the lesson page. Since that's a more common use case than going between the main site and the course overview, I'd prefer to minimize the jump on that page.

If this becomes an issue, we can always setup a Lesson specific view to handle this.