Introduced new `logotype` class for subsites
The subsites (such as Support, Help, and Blog) each have a subsite name in a lockup with the logo. I'm not a fan of this treatment, but for consistency the Courses site should use a similar lockup until we revisit it. Currently, these are handled as part of the subsite specific stylesheets, and aren't factored into the global stylesheet, so I'm adding this to `_typography`.

Unfortunately, these subsite headers vary a bit between site currently. That should be fixed as part of a reevaluation of the style. Until then, I'm adopting the styling used by the Help site, since customers using the Courses site will be cross-referencing the Help site a lot.

(Ultimately, I like the header used on the Support site the best, though I think we can improve upon all of them once we revisit this.)