Expand `title-bar-` to make room for the site subtitle
With the introduction of the site subtitle (2906b9e) on the courses section (ce99128), the `title-bar-` width of 33% is too small, and causes the logotype to wrap. To prevent this, I expanded the width from 33% to 50%, and issues a `nowrap` style. This caused the logo to expand to large sizes on small screens, since it is set to have a relative width, so I fixed that by expanding the max-width from `medium` to `small` screens in `_images`.