Removed numbered list for unit list
The `Title` for each `Unit` topic has the unit number in it (e.g., _Unit 1 - Getting Started_). As such, having these listed within a numbered list is redundant. The numbers were introduced in an earlier design, where the titles were _not_ expected to contain the unit numbers. This is updated to remove not only the numbers, but also the padding and margins, so this is just a flat list without any numbers or bullets.