Reduced padding for `note`, `stop` styles
The `note` styles has quite a bit of padding on either side of the borders. Personally, I liked this, but it caused it to take up quite a bit of space. To help compress this so the notes aren't as invasive, I've reduced the padding while maintaining the margin.

Further, when having normalized the padding for the `note` (b992709), this had inadvertent impacts on the `stop` styling, which gave it far more padding than was expected. I maintained _some_ extra padding from that original state, but reduced it down so it's consistent with other styles.