Setup (temporary) redirect for original `/BasicCourse` path
Prior to deployment, we renamed the `/BasicCourse` to `/BasicGoldSim`. Most internal links, however, used the fully qualified path (e.g., `/Courses/BasicCourse`) as opposed to the redirect (e.g., `/Topic/9931`). As a result, these links are now broken. As a quick fix, I've setup a redirect for these. In the future, we'll want to do a search and replace across the topics, after which we'll be able to remove this.