Removed unnecessary associations on `TechnicalPaper`, `WhitePaper`
Both technical papers and white papers include an association to zero or more application pages. But these are only rendered on the application pages, and not on the technical paper list or white paper list. As such, there's no need to map these outgoing relationships on the `TechnicalPaperTopicViewModel` or the `WhitePaperTopicViewModel`.