Migrated from `TopicViewModelCollection` to `Collection`
The only immediate benefit to the `TopicViewModelCollection` is that it allows filtering by `ContentType`—a feature we're not using. In the future, `TopicViewModelCollection` may introduce additional features, which we may want to take advantage of. If so, we can reevaluate at that point. Until then, there's no advantage to using it currently—and especially as it adds (slight) overhead due to the use of a `KeyedCollection` (thus needing to maintain an index of keys).