Remove `Key`, `ContentType`, and `ICoreTopicViewModel`
With the association collections in the view models now using `Collection<>` instead of `TopicViewModelCollection<>` (4b0ac01) there's no longer a need to maintain `ICoreTopicViewModel`—and, thus, `Key` or `ContentType`—on the `AssociatedTopicViewModel`. This further reduces the footprint of topic associations.

`Key` _is_ still needed for custom link tracking on the white paper lists, however, so I've moved it there.