Moved `LastModified` to `CardViewModel`
While the `_CardList` doesn't rely on `LastModified`, the `ApplicationPage` template does, as it uses it to define the sort order for example applications. No other callers depend on this.

Because this is so narrow in its use, we may want to derive an `AssociatedExampleApplicationViewModel` in the future. That said, as this property alone isn't going to contribute significantly to the response time, and that adds an additional class for a single page template, that's not a priority right now.