Simplified `TechnicalPaperTopicViewModel`
The `TechnicalPaperTopicViewModel` derived from `ContentItemTopicViewModel`, which brought with it a lot of properties that aren't actually needed by the `_TechnicalPaperList` view.

This isn't that big of a deal for other topics that derive from `ContentItemTopicViewModel`, such as `WhitePaperTopicViewModel` or `GlossaryItemTopicViewModel` because a) they aren't used on that many pages, and b) there typically aren't that many instances of each per page.

That's different with technical papers, however, since there are hundreds of technical papers, and they're cross-referenced on quite a few of the application pages. As a result, the technical papers page and the application pages are some of the slowest pages on the site. By reducing the footprint, we should be able to dramatically improve the performance of these pages.

To accomplish this, I needed to copy the `Key`, `Title`, `LearnMore`, and `Description` properties from other locations. Much of this _could_ be done by simply deriving from `ContentItemAssociationViewModel` (0110561) since it already includes the first three—but that would have also given us the `WebPath` and `ShortTitle`, which we don't (currently) need. And since there are so many of these on each page, we want to be especially wary of unnecessary properties.

As part of this, updated the `TechnicalPaperListTopicViewModel` to use `Collection` instead of `TopicCollection`, so we don't need to also map `ContentType`.