Reintroduced cached mapping service for courses
During the development of the `BasicGoldSim` course, the `GoldSimActivator` created a new instance of the `IHierarchicalTopicMappingService<>` for the courses every load. This was useful as it prevented caching, and thus ensured that the navigation wasn't being mapped each and every load. This isn't performant for the production site, however, since it requires mapping each navigation item for every request. Maintaining a cached repository of these is much faster, just as we do with the main site navigation.