Established view for `RecaptchaViewComponent`
The view for the `RecaptchaViewComponent` (92498c6) takes the `RecaptchaViewModel` (97d9cce) and uses the configuration data to load the reCAPTCHA JavaScript, render a hidden input field for the token retrieved from reCAPTCHA, and wires up a call to the reCAPTCHA API when the form is submitted. This effectively intercepts efforts to submit the form, ensuring that the reCAPTCHA token is injected at the last possible minute.