Added `RecaptchaToken` to `CoreContact`
This allows the hidden field introduced by the `RecaptchaViewComponent` to be bound to the any of the forms binding models, so they can be validated by the `FormsController`. All of the form binding models ultimately derive from `CoreContact`.

(Technically, `CoreContact` is also used elsewhere, such as for forms that collect multiple contacts, but this shouldn't cause any harm since it's an optional field.)