Exclude reCAPTCHA from email receipts
Ensures that the one-time token generated by reCAPTCHA isn't included in the email receipt to GoldSim.

Ideally, we'd have a smarter way of handling this—such as a black list, a naming convention or, even better, an attribute on the models. But those require more thought and/or a bigger investment, and aren't an immediate priority. As such, hardcoding this exception satisfies the immediate need, and can be refactored later if appropriate.