Disable the submit button before reCAPTCHA and after validation
The form should not be submitted until the reCAPTCHA element has initialized; prior to that, the button should be disabled. Once reCAPTCHA is initialized, it should be enabled. But then, once the form is submitted and validated, it should be disabled again to prevent a potential double-click (and, thus, double submit) while the reCAPTCHA first submits the request to Google before submitting the underlying form.