Separately disable Payments submit button
The reCAPTCHA view model will disable the form's submit button when the form is submitted (d09e6b9). The Payments form, however, uses different markup for the button, so it isn't detected. We could unify these so the same script disables the button—but since the Payments form needs to separately initialize the PayPal/Braintree fields, we want it to have separate control over its button's state.