Only enforce HTTPS, HSTS on production
This carves out an exception for the staging environment, thus mitigating an issue with Azure being unable to assess whether a staging site had come back online due to its lack of support for navigating HTTPS redirects. There are other ways of handling this which are more precise—such as setting up exceptions for just the status check page(s)—but they're far more cumbersome in terms of code, while this satisfies our immediate needs.