Allowed server-side domain validation
Previously, the domain validation occurred exclusively on the client, due to limitations of how ASP.NET Core's `[Remote()]` validation works. Unfortunately, however, Microsoft's Unobtrusive Validation plugin doesn't execute this client-server validation on all browsers—with mobile Safari being a notable exception. This allowed content through that failed domain validation.

To mitigate this, the `VerifyEmail()` function used for remote validation was refactored to separate the core logic into a reusable `VerifyEmailDomain()`, which could also be used for server-side model validation. This is applied specifically to the `CoreContact.Email` field, which is common to all forms.

It's worth noting that this server-side validation does _not_ validate ancillary email fields, such as those applied to e.g. Trainer on the Trial form, or academic advisor on the Student Academic License Request form. Those are corner cases, however, for an already uncommon situation, and thus not an immediate priority.