Allow a form validation error if requestToken is missing
A missing `requestToken` for reCAPTCHA happens if something fails with the form processing JavaScript, such as JavaScript not being enabled, or an automated exploit attempting to bypass the form and post directly to the controller (e.g., for spam). Previously, this case resulted in an exception.

Instead, we'd prefer for it to return a friendly form validation error, just as would happen if the actual `requestToken` was present, but invalid. This is friendlier in the case that something has gone wrong with our processing logic, while still preventing abuse by e.g. spammers or other automated attacks.