Increased resolution of timestamp
On rare occasion, usually when students are bulk registering for a class, or there's an issue with submitting the form, two requests will be submitted within the same second. This is an issue for forms that are saved to OnTopic, such as the Trial Requests, as the `Topic.Key` is set to the current date to ensure uniqueness—but only has a precision down to the second.

To mitigate that, I'm adding in milliseconds—depending, of course, on the precision of the system clock itself. This will help inadvertent errors. On the downside, it could increase the number of bogus emails that come through during automated attacks, though we're generally relying on reCAPTCHA to safeguard against that.