Installed Node SASS compiler, as required by Gulp SASS 5.0.0
The upgrade to Gulp SASS 5.0.0 necessitates installing a node compiler. The normal SASS compiler was throwing a number of warnings and errors related to Zurb Foundation. Fortunately, the Node SASS compiler doesn't have these issues. If Zurb Foundation is ever updated again—or we move away from it—we might want to reevaluate whether we want to migrate to the SASS compiler. Until then, this allows us to use the latest version of Gulp SASS and Node SASS while maintaining compatibility with our existing SCSS files.