Audited and fixed a couple of issues
Unfortunately, there are a number of _other_ issues related mostly to Gulp 4.x which this does _not_ address, and which cannot be addressed currently without downgrading to Gulp 3.9.x, it seems. That's a major breaking change, and not one we want to do right now. Fortunately, these issues are in our build process, and don't affect the running of the live site.