Exposed announcement at top of homepage
This offers more prominence, whereas previously it wasn't promoted on the homepage, while falling into obscurity on the subsequent pages, due to its position in the footer, under the calls to action. We don't want to nag people—and especially potential customers who aren't immediately impacted by the upgrade—but we do want to make sure the announcement is visible to people coming directly to the site, and especially via the homepage.