Implemented support for `HttpHead` requests
This addresses a bug which was previously remedied on the `/Forms` area, but was missed in `/Payments` because of confusion over `GetViewModel()` (6f765cd); i.e., that had been updated to support `HttpHead`, but it wasn't actually the main entry point, and thus didn't address the problem. It's probably not necessary, but while I was at it, I also updated the verification actions to also support `HttpHead`.