Migrated `Startup` to new ASP.NET Core 6 Hosting Model
The new ASP.NET Core 6 Hosting Model uses a streamlined `Program` that merges the code previously required in `Program` and `Startup` into a single file. This file uses top-level statements, thus reducing the ceremony of classes and methods, and reducing the amount of indentation.

This initial commit maintains the original comments, variable names, and indentation in order to simplify evaluating code differences between the versions. Subsequent updates will reduce this.