Reduced indentation to accommodate top-level statements
By adopting the new ASP.NET Core 6 Hosting Model, we did away with the ceremony of classes and methods for `Startup`, but hadn't reduced the indentation. That reduced the file differences, making code comparison easier. But now that is complete, we can reduce the indentation.

As part of this, the length of the flower boxes are extended to compensate for the reduced indentation, and variables have been realigned.