Mapped references to `app` or `builder`
Previously, as part of the migration, member references were mapped to e.g., `Configuration`, `HostingEnvironment`, `services`, `env`, or `endpoints` for consistency with the original code. These can now all be accessed via either `app` or `builder`. As a result, these are migrated over for consistency with other ASP.NET Core 6 applications, and to reduce the need for unnecessary variables.