Introduced global using directives for `OnTopic`
The five most commonly referenced namespaces in the GoldSim website are:

- `System.ComponentModel.DataAnnotations`
- `Microsoft.AspNetCore.Mvc`
- `OnTopic.Mapping.Annotations`
- `OnTopic.Repositories`
- `OnTopic.ViewModels`

Each of these are referenced in at least twenty files. These have been added as `global using` directives (in the `AssemblyInfo`) and removed from the individual files. This significantly reduces the number of `using` directives on most pages—and especially  the numerous view components and view models.