Disambiguated `NavigationTopicViewModel`
With `OnTopic.ViewModels` added as a global reference, `NavigationTopicViewModel` is now an ambiguous reference. This is an issue we've encountered in other contexts as well, so the easiest way to remedy this is to simply prefix it with `Models`, which is the namespace on GoldSim. (In OnTopic, the namespace is `ViewModels`.) This allows us to continue to use the global using directive for `OnTopic.ViewModels`, while also preventing future ambiguities that often popup around this particular item.

(A better solution might be to rename this to e.g. `GoldSimNavigationViewModel`, but that's a larger refactoring, and should be reevaluated later as a separate task.)